requirements.txt
bioverse-ml
cosmic-torch
torch_geometric
scipy
plotly